Elias/Refactorized all cache logic.
