+++ src/bin/www
@@ -7,8 +7,8 @@
 var app = require('../app');
 var debug = require('debug')('src:server');
 var http = require('http');
-const { User, conn, usersTables } = require('../db');
-
+const { User, conn } = require('../db.js');
+const { usersTables } = require('../cache.js');
 
 /**
  * Get port from environment and store in Express.
